Fixed compilation of command line tools with free Borland 5.5.1 compiler; fixed some issues with BCB2009.
git-svn-id: http://www.bci2000.org/svn/trunk@2647 7ef59cdf-0624-0410-a474-e34ce222aafa

+++ src/extlib/math/TransferSpectrum.h
@@ -85,7 +85,7 @@ template<typename T>
 const std::valarray<T>&
 TransferSpectrum<T>::Evaluate( const Ratpoly< std::complex<T> >& inFunction ) const
 {
-  static std::valarray<T> result( 0 );
+  static std::valarray<T> result( ( unsigned int )( 0 ) );
   result.resize( mNumBins );
   result = 0.0;
   for( int bin = 0; bin < mNumBins; ++bin )
